fix css insert plugin
